add Map overlays feature folder
