+++ packages/tupaia-web/src/features/Map/Map.tsx
@@ -5,19 +5,14 @@
 
 import React, { useState } from 'react';
 import styled from 'styled-components';
+import { useParams } from 'react-router-dom';
 import { TileLayer, LeafletMap, ZoomControl, TilePicker } from '@tupaia/ui-map-components';
-import {
-  TRANSPARENT_BLACK,
-  TILE_SETS,
-  MOBILE_BREAKPOINT,
-  URL_SEARCH_PARAMS,
-} from '../../constants';
+import { TRANSPARENT_BLACK, TILE_SETS, MOBILE_BREAKPOINT } from '../../constants';
 import { MapWatermark } from './MapWatermark';
 import { MapLegend } from './MapLegend';
-import { PolygonLayer } from './PolygonLayer';
+import { MapOverlays } from '../MapOverlays';
 import { MapOverlaySelector } from './MapOverlaySelector';
-import { useEntities, useLegacyMapOverlay } from '../../api/queries';
-import { useParams, useSearchParams } from 'react-router-dom';
+import { useEntities } from '../../api/queries';
 
 const MapContainer = styled.div`
   height: 100%;
@@ -71,23 +66,12 @@ const TilePickerWrapper = styled.div`
   }
 `;
 
-const ENTITY_FIELDS = ['parent_code', 'code', 'name', 'type', 'bounds', 'region'];
-
 export const Map = () => {
   const { projectCode, entityCode } = useParams();
   const [activeTileSet, setActiveTileSet] = useState(TILE_SETS[0]);
 
-  const [urlSearchParams] = useSearchParams();
-  const overlayCode = urlSearchParams.get(URL_SEARCH_PARAMS.OVERLAY);
-  const { data: mapOverlayData } = useLegacyMapOverlay(overlayCode, {
-    params: {
-      organisationUnitCode: entityCode,
-      projectCode,
-      shouldShowAllParentCountryResults: true,
-    },
-  });
   const { data: entityData } = useEntities(projectCode, entityCode, {
-    params: { fields: ENTITY_FIELDS },
+    params: { fields: ['parent_code', 'code', 'name', 'type', 'bounds', 'region'] },
   });
 
   const onTileSetChange = (tileSetKey: string) => {
@@ -98,7 +82,7 @@ export const Map = () => {
     <MapContainer>
       <StyledMap bounds={entityData?.bounds} shouldSnapToPosition>
         <TileLayer tileSetUrl={activeTileSet.url} showAttribution={false} />
-        <PolygonLayer entity={entityData} mapOverlayData={mapOverlayData} />
+        <MapOverlays />
         <ZoomControl position="bottomright" />
         <MapLegend />
         <MapWatermark />
