+++ packages/tupaia-web/src/features/MapOverlays/MarkerLayer.tsx
@@ -1,0 +1,10 @@
+/**
+ * Tupaia
+ * Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+
+import React from 'react';
+
+export const MarkerLayer = () => {
+  return <></>;
+};
