+++ packages/tupaia-web/src/features/MapOverlays/index.ts
@@ -1,0 +1,5 @@
+/*
+ * Tupaia
+ *  Copyright (c) 2017 - 2023 Beyond Essential Systems Pty Ltd
+ */
+export { MapOverlays } from './MapOverlays.tsx';
